requirements.txt
Django==4.2.8
djangorestframework==3.16.0
django-oauth-toolkit==3.0.1
pytest==8.4.1
pytest-django==4.11.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
drf-spectacular==0.28.0